代码同步

+++ rosen/modules/composer/hdi_backend/include/hdi_layer_info.h
@@ -394,12 +394,12 @@ public:
         presentTimestamp_ = timestamp;
     }
 
-    int32_t GetSdrNit() const
+    float GetSdrNit() const
     {
         return sdrNit_;
     }
 
-    int32_t GetDisplayNit() const
+    float GetDisplayNit() const
     {
         return displayNit_;
     }
@@ -409,12 +409,12 @@ public:
         return brightnessRatio_;
     }
 
-    void SetSdrNit(int32_t sdrNit)
+    void SetSdrNit(float sdrNit)
     {
         sdrNit_ = sdrNit;
     }
 
-    void SetDisplayNit(int32_t displayNit)
+    void SetDisplayNit(float displayNit)
     {
         displayNit_ = displayNit;
     }
